dynamic server tag support - basic unit converter with div using exec="server" works
- basic synchronization for tag objects - events from client/server and innerHTML from server to client

- javascript code-gen fix - look at base types to pick up JSSettings dependencies

- start layer for get file processor to fix sctjs in IDE

+++ system/src/sc/lang/html/Event.java
@@ -11,7 +11,7 @@ public class Event {
    public Element currentTarget;
    public Element currentTag;
    public Element target;
-   public long timeStamp;
+   public double timeStamp;
 
    // TODO: these are stubs for the client right now since there's no real event delivery on the server
    public void preventDefault() {
